Add policies for the Super Admin in Nova and update some UI stuff
In order to control what user Roles have access to certain resources
this commit creates a policy for all the resources and only grants
access to the `super_administrator` role. This commit also updates all
the Nova UI stuff to be the Napoleon Youth Basketball style.  This
commit also updates the nova route from `/zadmin` to `/admin`.

[902951390108168](https://app.asana.com/0/0/902951390108168/f)